requirements.txt
numba
pandas
numpy
matplotlib
numpyro
jax
jaxlib
torch
torchvision
pyro-ppl